requirements.txt
# Core dependencies
fastapi==0.128.0
uvicorn[standard]==0.32.1

# Document processing
python-docx==0.8.11
pypdf==4.0.1

# LLM & AI
langchain==0.1.17
langchain-core>=0.1.48,<0.2.0
langchain-groq==0.1.5
groq==0.37.1

# Utilities
python-dotenv==1.0.0
pydantic>=2.7.0
requests==2.31.0

# Development (optional)
pytest==7.4.0
black==23.12.0
flake8==6.1.0
